LayoutTests/http/tests/websocket/tests/hybi/close-on-unload-and-force-gc.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test if Web Socket is closed on unload and not crashed if garbage collected");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

function gc()
{
    if (window.GCController)
        return GCController.collect();

    for (var i = 0; i < 10000; i++) { // > force garbage collection (FF requires about 9K allocations before a collect)
        var s = new String("abc");
    }
};

var frameDiv;
var closedSocket;

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/close-on-unload");
ws.onopen = function()
{
    debug("PASS ws on master document is ready.");
    frameDiv = document.createElement("iframe");
    frameDiv.src = "resources/close-on-unload-iframe.html";
    document.body.appendChild(frameDiv);
    debug("PASS insert a iframe, where open ws called 'socket1'");
};
ws.onmessage = function(evt)
{
    closedSocket = evt.data;
    ws.close();
};
ws.onclose = function()
{
    shouldBe("closedSocket", '"socket1"');
    finishJSTest();
};

document.iframeReady = function()
{
    debug("PASS 'socket1' is sent to the server. unload the iframe and force garbage collection. expect receiving 'socket1' on ws and no crash...");
    document.body.removeChild(frameDiv);
    gc();
};

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
